Updated insatll instrctions and added dependency to argparse

+++ setup.py
@@ -68,15 +68,20 @@ def _install_bundle(install_dir, base=None):
         dirpath = tempfile.mkdtemp()
 
     target = "%s/%s" %(dirpath, file_name)
-    if not os.path.exists(target):
-        print "Downloading %s bundle from %s to %s" % (type, base_url, target)
-        urllib.urlretrieve (base_url, target)
+    keep = False
+    if os.path.exists(file_name):
+        target = file_name
+        keep = True
+    else:
+        if not os.path.exists(target):
+            print "Downloading %s bundle from %s to %s" % (type, base_url, target)
+            urllib.urlretrieve (base_url, target)
 
-    tar = subprocess.Popen("tar xzvf %s --exclude \"._*\"" % (target), shell=True, cwd=dirpath)
+    tar = subprocess.Popen("tar xzvf %s --exclude \"._*\"" % (os.path.abspath(target)), shell=True, cwd=dirpath)
     if tar.wait() != 0:
         print "Error while extracting gem bundle"
         exit(1)
-    if base is None:
+    if base is None and not keep:
         os.remove(target)
 
     bins=[x for x in os.listdir(dirpath)]
@@ -209,6 +214,7 @@ https://github.com/gemtools/gemtools
           'Programming Language :: C',
           'Topic :: Scientific/Engineering :: Bio-Informatics',
         ],
+        install_requires = ["argparse"],
         entry_points = {
             'console_scripts': [
                 'gem-rnaseq-pipeline = gem.production.gtex_rnaseq_pipeline:main'
